backend/requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.0.0
scikit-image==0.19.3
opencv-python==4.5.5.62
gunicorn==21.2.0